Duplicated Taskfile.yml
Problem:
There is duplicated logic in 2 Taskfiles - one
for inside the container and one for outside.

Solution:
Test the $USER variable for accepted values from a vscode or
builder container and run cmds without 'docker run' prefixed
o found.
Additionally pre-install nodejs for the pyright type-checker so
that pyright checks complete quickly.

+++ scripts/builder.sh
@@ -7,6 +7,12 @@
 #     ./scripts/builder.sh /bin/bash   # for shell
 #     ./scripts/builder.sh             # enters python REPL
 
+if [ "$USER" = "builder" -o "$USER" = "vscode" ]
+then
+    "$@"
+    exit 0
+fi
+
 docker run \
     --rm -it \
     -v $(pwd):/home/builder \
